+++ lib/module/mixins.rb
@@ -1,1 +1,3 @@
 require 'module/mixins/version'
+require 'module/mixins/including'
+require 'module/mixins/included'