+++ lib/module/mixins/included.rb
@@ -1,0 +1,8 @@
+module Module::Mixins::Included
+  attr_accessor :classes_including
+
+  def included(base)
+    self.classes_including ||= []
+    self.classes_including << base
+  end
+end