+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ### 0.1.0 (Next)
 
-* Initial public release - [@dblock](https://github.com/dblock).
+* Initial public release with `Module::Mixins::Included` and `Module::Mixins::Including` - [@dblock](https://github.com/dblock).